+++ Classes/Interfaces/IPriority.cs
@@ -1,0 +1,6 @@
+﻿namespace CloverAPI.Classes.Interfaces;
+
+internal interface IPriority
+{
+    int Priority { get; }
+}