+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 0.1.1
 - Integrated settings by pharmacomaniac a.k.a. zapybibby
+- Added support for audio and texture overrides
+- Most string ids are now case-insensitive when checking for matches
 - Added a patch to fix an error being logged when the platform kind is accessed before initializing Master
 - Added `GameUtils.GameReady` property and corresponding `OnGameReady` event to signal when everything is initialized- Made the 'Modded' on-screen text darker and less obtrusive
 - Modded Charms now should show up in the terminal